requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.2
chardet==5.2.0
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
html5lib==1.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
iopath==0.1.10
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
kiwisolver==1.4.8
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-openai==0.3.9
langchain-postgres==0.0.13
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.15
layoutparser==0.3.4
lxml==5.3.1
Markdown==3.7
marshmallow==3.26.1
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.2.0
narwhals==1.31.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.21.0
openai==1.66.3
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pgvector==0.3.6
pi_heif==0.22.0
pikepdf==9.5.2
pillow==11.1.0
portalocker==3.1.1
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psycopg==3.2.6
psycopg-pool==3.2.6
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pypandoc==1.15
pyparsing==3.2.1
pypdf==5.4.0
pypdfium2==4.30.1
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.1
RapidFuzz==3.12.2
redis==5.2.1
regex==2024.11.6
rsa==4.9
safetensors==0.5.3
scipy==1.15.2
smmap==5.0.2
SQLAlchemy==2.0.39
streamlit==1.43.2
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.1
unstructured==0.16.10
unstructured-client==0.31.1
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.15
watchdog==6.0.0
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.2
yarl==1.18.3
zstandard==0.23.0
faiss-cpu==1.11.0
accelerate==1.7.0
timm==1.0.15
